pyproject.toml
[tool.poetry]
name = "Auguri-MusikalischesWurfelspiel"
version = "1.0"
description = "Script per generare un video di auguri musicali basato sul gioco compositivo Musikalisches Würfelspiel"
authors = ["Burlesco70 <mario.nardi.70@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
